Removes duplicated data augmentation code; Remove tfrecord utility routines from tfrecord.py and train.py and move to tfrecord_utils.py; Update the training notebook to load tfrecord data.
